+++ src/azure-cli/azure/cli/command_modules/vm/tests/latest/test_vm_commands.py
@@ -1199,16 +1199,17 @@ class VMCreateNoneOptionsTest(ScenarioTest):  # pylint: disable=too-many-instanc
 
 class VMCreateMonitorTest(ScenarioTest):
 
-    @ResourceGroupPreparer(name_prefix='cli_test_vm_create_with_monitor', location='centralus')
+    @ResourceGroupPreparer(name_prefix='cli_test_vm_create_with_monitor', location='eastus')
     def test_vm_create_with_monitor(self, resource_group):
 
         self.kwargs.update({
             'vm': 'monitorvm',
-            'workspace': 'vmlogworkspace20191009',
+            'workspace': self.create_random_name('cliworkspace', 20),
             'rg': resource_group
         })
 
         self.cmd('vm create -n {vm} -g {rg} --image UbuntuLTS --workspace {workspace}')
+        self.cmd('vm monitor log show -n {vm} -g {rg} -q "Perf | limit 10"')
 
 
 class VMBootDiagnostics(ScenarioTest):